move entities to virtual-entities, user repository and small changes
