+++ server/modules/analysis/api/entitiesApi.js
@@ -6,18 +6,22 @@ import * as AnalysisService from '../service'
 export const init = (app) => {
   // ====== GET - virtual_entities
 
-  app.get('/survey/:surveyId/entities/', AuthMiddleware.requireRecordAnalysisPermission, async (req, res, next) => {
-    try {
-      const { cycle, surveyId, offset, limit } = Request.getParams(req)
-      const list = await AnalysisService.fetchVirtualEntities({ surveyId, cycle, offset, limit })
-      res.json({ list })
-    } catch (error) {
-      next(error)
+  app.get(
+    '/survey/:surveyId/virtual-entities/',
+    AuthMiddleware.requireRecordAnalysisPermission,
+    async (req, res, next) => {
+      try {
+        const { cycle, surveyId, offset, limit } = Request.getParams(req)
+        const list = await AnalysisService.fetchVirtualEntities({ surveyId, cycle, offset, limit })
+        res.json({ list })
+      } catch (error) {
+        next(error)
+      }
     }
-  })
+  )
 
   app.get(
-    '/survey/:surveyId/entities/count',
+    '/survey/:surveyId/virtual-entities/count',
     AuthMiddleware.requireRecordAnalysisPermission,
     async (req, res, next) => {
       try {
