--- server/modules/analysis/manager/entity/index.js
@@ -1,4 +1,0 @@
-import * as EntityRepository from '../../repository/entity'
-
-// ====== READ
-export const { fetchVirtualEntities, countVirtualEntities } = EntityRepository
