--- server/modules/analysis/repository/entity/index.js
@@ -1,1 +1,0 @@
-export { fetchVirtualEntities, countVirtualEntities } from './read'
