+++ server/modules/nodeDef/repository/nodeDefRepository.js
@@ -3,6 +3,8 @@ import * as R from 'ramda'
 import { db } from '@server/db/db'
 import * as DbUtils from '@server/db/dbUtils'
 import * as NodeDef from '@core/survey/nodeDef'
+import * as DB from '@server/db'
+import { TableNodeDef } from '@openforis/arena-server'
 import {
   getSurveyDBSchema,
   dbTransformCallback as dbTransformCallbackCommon,
@@ -343,4 +345,53 @@ export const deleteNodeDefsValidationMessageLabels = async (surveyId, langs, cli
     WHERE
       e.uuid = n.uuid
   `)
+}
+
+/**
+ * Fetches all virtual entities.
+ *
+ * @param {!object} params - The query parameters.
+ * @param {!string} params.surveyId - The survey id.
+ * @param {number} [params.offset=0] - The select query offset.
+ * @param {number} [params.limit=null] - The select query limit.
+ * @param {BaseProtocol} [client=db] - The database client.
+ *
+ * @returns {Promise<any[]>} - The result promise.
+ */
+export const fetchVirtualEntities = async (params, client = DB.client) => {
+  const { surveyId, offset = 0, limit = null } = params
+
+  const tableNodeDef = new TableNodeDef(surveyId)
+
+  return client.map(
+    ` 
+    select
+  _nd.*,
+  _nd.props || _nd.props_draft as props,
+  _ndp.props || _ndp.props_draft as source_props
+  from
+  ${tableNodeDef.nameQualified} as _nd
+  Left join ${tableNodeDef.nameQualified} as _ndp
+  on _ndp.uuid = _nd.parent_uuid
+  where _nd.virtual = TRUE and _nd.deleted = FALSE
+  LIMIT ${limit || 'ALL'}
+  OFFSET ${offset}`,
+    [],
+    dbTransformCallback
+  )
+}
+
+/**
+ * Count virtual entities.
+ *
+ * @param {!object} params - The query parameters.
+ * @param {!string} params.surveyId - The survey id.
+ * @param {BaseProtocol} [client=db] - The database client.
+ *
+ * @returns {Promise<number>} - The result promise.
+ */
+export const countVirtualEntities = async (params, client = DB.client) => {
+  const { surveyId } = params
+  const tableNodeDef = new TableNodeDef(surveyId)
+  return client.one(`select count(*) from ${tableNodeDef.nameAliased} where virtual = TRUE and deleted = FALSE`)
 }
