+++ webapp/components/survey/NodeDefDetails/store/actions/useSaveEdits.js
@@ -45,9 +45,9 @@ export const useSaveEdits = ({ setState }) => {
         )
       }
       let chainNodeDef = null
-      if (NodeDef.isAnalysis(nodeDefUpdated)) {
+      if (NodeDef.isAnalysis(nodeDefUpdated) && NodeDef.isAttribute(nodeDef)) {
         chainNodeDef = {
-          chainUuid: chain?.uuid,
+          chainUuid: chain.uuid,
           nodeDefUuid: NodeDef.getUuid(nodeDefUpdated),
           props: { active: true },
           uuid: uuidv4(),
